requirements.txt
slackclient==2.9.0
slackeventsapi==2.2.1
flask==1.1.0
python-dotenv==0.14.0
flask-sqlalchemy==2.4.1
sqlalchemy==1.3.17
sqlalchemy-utils==0.36.5
flask-marshmallow==0.12.0
pymysql==0.9.3
